No longer relies on formation IDs >=500
Previously the loader assumed IDs in the load string were formations if >=500, and upgrades otherwise. This constraint is not however enforced in the builder, leading to failures with lists where this constraint was not observed.

This fix removes this assumption and simply tests whether the ID identifies a formation or not. If it does, it is loaded regardless of whether it's >= 500.