Update SM_scions_NETEA.json
Updated Typhoon's cost.
Updated Snipers' cost. Removed unlimited Snipers per formation.
Updated Whirlwinds' cost.
Updated pair of Land Raiders or Crusaders' cost.
Changed the structure so a maximum of four extra LR of any kind could be purchased at most, while keeping the pair of extra LR or Crusader at a different cost but not on the base ones, and breaking the least possible previously saved forces. With the previous one, 4 extra LRs plus 4 more extra Helios per formation could be purchased. It is still not perfect as it is not possible to purchase odd numbers of Helios when purchasing the maximum of extra LRs.