Update SM_codex_NETEA.json
Added choice for Dreadnought and Razorback's weapon configurations
Added maximum Razorbacks to keep in line with the others.
Reworked Snipers so they are chosen as characters like intended.
Added ''plus transport''. Kept Rhinos and Drop Pods as lines for those that play it as part of force selection instead on how it is in the special rule.
Changed 33% for one third.