Update SM_codex_NETEA.json
Added choice for Dreadnought and Razorback's weapon configurations.
Added maximum Razorbacks per formation to keep in line with the other upgrades.
Reworked Snipers so they are chosen as characters like intended.
Added ''plus transport''. Kept Rhinos and Drop Pods as lines for those that houserule it as part of force selection.
Changed 33% for one third.